+++ README.md
@@ -1,8 +1,3 @@
 # FailyV Unofficial bank app
 
 ![Project screenshot](/docs/preview.png)
-
-## TODO
-
-- [ ] log list filters
-- [ ] Form automatic button states
