+++ src/scripts/classes/notification-list.js
