+++ src/scripts/classes/theme.js
@@ -1,0 +1,3 @@
+class Theme {
+
+}
