+++ src/scripts/classes/view-switcher.js
@@ -1,0 +1,12 @@
+export class ViewSwitcher {
+  constructor(initialView) {
+    this.activeView = initialView;
+    this.activeView.activate();
+  }
+
+  switch(view) {
+    this.activeView.deactivate();
+    this.activeView = view;
+    this.activeView.activate();
+  }
+}
