+++ src/scripts/classes/view.js
@@ -1,23 +1,10 @@
-export class ViewSwitcher {
-  constructor(initialView) {
-    this.activeView = initialView;
-    this.activeView.activate();
-  }
-
-  switch(view) {
-    this.activeView.deactivate();
-    this.activeView = view;
-    this.activeView.activate();
-  }
-}
-
 export class View {
   constructor(
     viewElement,
     filterList,
     logList,
     favoriteAccountList
-    ) {
+  ) {
     this.viewElement = viewElement;
     this.filterList = filterList;
     this.logList = logList;
