+++ src/styles/app.css
@@ -1,0 +1,68 @@
+@keyframes spin {
+    0% {
+        transform: rotate(0deg);
+    }
+
+    100% {
+        transform: rotate(360deg);
+    }
+}
+
+.app {
+    box-sizing: border-box;
+    position: relative;
+    display: flex;
+    width: 1224px;
+    height: 720px;
+    margin: auto;
+    background-color: var(--color-bg);
+    border: 1px solid var(--color-border-on-bg);
+    overflow: hidden;
+}
+
+.app--loading > .app__spinner {
+    display: block;
+}
+
+.app__spinner {
+    display: none;
+    box-sizing: border-box;
+    width: 72px;
+    height: 72px;
+    border-radius: 9999px;
+    border: 8px solid var(--color-toggle-button-track);
+    animation: spin 2s linear infinite;
+    margin: auto;
+}
+
+.maze .app__spinner {
+    border-top: 8px solid var(--color-maze);
+}
+
+.fleeca .app__spinner {
+    border-top: 8px solid var(--color-fleeca);
+}
+
+.app__container {
+    box-sizing: border-box;
+    display: none;
+    width: 100%;
+    height: 100%;
+}
+
+.app--loaded > .app__container {
+    display: block;
+}
+
+.app__notification-list {
+    box-sizing: border-box;
+    position: absolute;
+    z-index: 100;
+    bottom: 0;
+    right: 0;
+    width: 400px;
+    height: auto;
+    margin: 0;
+    padding: 0 24px 0 0;
+    list-style: none;
+}
