+++ src/styles/favorite-account.css
@@ -1,0 +1,132 @@
+.favorite-account {
+    box-sizing: border-box;
+    width: 100%;
+    height: auto;
+    border-bottom: 1px solid var(--color-border);
+}
+
+.favorite-account--clickable {
+    cursor: pointer;
+}
+
+.favorite-account:last-child {
+    border: none;
+}
+
+.favorite-account__section {
+    box-sizing: border-box;
+    width: auto;
+    height: auto;
+    padding-top: 10px;
+    padding-bottom: 10px;
+}
+
+.favorite-account__section--right {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+}
+
+.favorite-account__accordion {
+    box-sizing: border-box;
+    position: relative;
+    display: block;
+    width: auto;
+}
+
+.favorite-account__name {
+    box-sizing: border-box;
+    display: block;
+    font-size: 16px;
+    font-weight: 700;
+    color: var(--color-text);
+    line-height: 20px;
+    margin: 0;
+}
+
+.favorite-account__name--margin {
+    margin-bottom: 6px;
+}
+
+.favorite-account__number {
+    box-sizing: border-box;
+    display: block;
+    margin: 0;
+    font-size: 14px;
+    font-weight: 600;
+    color: var(--color-text-alt);
+    line-height: 1;
+}
+
+.favorite-account__icon-button {
+    box-sizing: border-box;
+    width: 36px;
+    height: 36px;
+    list-style: none;
+    background-color: transparent;
+    background-image: url("/src/images/delete.svg");
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: 16px;
+    cursor: pointer;
+    margin-right: -10px;
+}
+
+.favorite-account__accordion[open] > .favorite-account__icon-button {
+    background-image: url("./images/close.svg");
+}
+
+.favorite-account__icon-button:hover {
+    background-color: var(--color-icon-bg);
+}
+
+.favorite-account__icon-button:focus-visible {
+    outline: 2px solid var(--color-down);
+}
+
+.favorite-account__text-button {
+    box-sizing: border-box;
+    display: block;
+    position: absolute;
+    top: 0;
+    left: -8px;
+    transform: translateX(-100%);
+    width: auto;
+    height: 36px;
+    padding-left: 16px;
+    padding-right: 16px;
+    border: none;
+    background-color: var(--color-down);
+    font-family: inherit;
+    font-size: 10px;
+    font-weight: 700;
+    text-transform: uppercase;
+    letter-spacing: 0.6px;
+    color: #fff;
+    cursor: pointer;
+}
+
+.favorite-account__text-button:focus-visible {
+    outline-offset: 3px;
+    outline: 2px solid var(--color-down);
+}
+
+.favorite-account__paste-icon {
+    box-sizing: border-box;
+    width: 48px;
+    height: 48px;
+    background-color: transparent;
+    background-image: url("/src/images/paste-form.svg");
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: 24px;
+    margin-right: -12px;
+}
+
+.maze .favorite-account--clickable:hover .favorite-account__name {
+    color: var(--color-maze);
+}
+
+.fleeca .favorite-account--clickable:hover .favorite-account__name {
+    color: var(--color-fleeca);
+}
