+++ src/styles/filter.css
@@ -1,0 +1,22 @@
+.filter {
+    display: block;
+    border: none;
+    background-color: transparent;
+    font-family: inherit;
+    font-size: 16px;
+    font-weight: 600;
+    line-height: 24px;
+    width: 100%;
+    text-align: left;
+    cursor: pointer;
+    color: var(--color-text-alt);
+}
+
+.filter:hover {
+    background-color: var(--color-border);
+    color: var(--color-text);
+}
+
+.filter--active {
+    color: var(--color-border);
+}
