+++ src/styles/paper.css
@@ -1,0 +1,52 @@
+.paper {
+    position: relative;
+    box-sizing: border-box;
+    width: 100%;
+    height: 100%;
+    background-color: var(--color-fg);
+}
+
+.paper__frame {
+    box-sizing: border-box;
+    position: absolute;
+    z-index: 50;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    overflow-y: auto;
+    padding-left: 16px;
+    padding-right: 16px;
+}
+
+.paper__frame::-webkit-scrollbar {
+    width: 10px;
+}
+
+.paper__frame::-webkit-scrollbar-thumb {
+    background-color: var(--color-scrollbar-thumb);
+}
+
+.paper__frame::-webkit-scrollbar-track {
+    background-color: transparent;
+}
+
+.paper__table {
+    box-sizing: border-box;
+    width: 100%;
+    height: auto;
+    border-collapse: collapse;
+}
+
+.paper__button {
+    box-sizing: border-box;
+    display: none;
+    position: absolute;
+    z-index: 60;
+    bottom: 16px;
+    right: 16px;
+}
+
+.paper:hover > .paper__button {
+    display: flex;
+}
