+++ src/styles/section.css
@@ -1,0 +1,79 @@
+.section {
+    box-sizing: border-box;
+    width: 100%;
+    height: auto;
+}
+
+.section--column {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+}
+
+.section__row {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.section__title {
+    box-sizing: border-box;
+    display: block;
+    font-size: 12px;
+    font-weight: 600;
+    text-transform: uppercase;
+    letter-spacing: 0.6px;
+    color: var(--color-text-alt);
+    line-height: 1;
+    margin: 0 0 8px 0;
+}
+
+.section__number {
+    font-size: 24px;
+    font-weight: 700;
+    color: var(--color-text);
+    letter-spacing: 0.8px;
+    line-height: 1;
+    margin: 0;
+}
+
+.section__balance {
+    font-size: 24px;
+    font-weight: 700;
+    color: var(--color-up);
+    letter-spacing: 0.8px;
+    line-height: 1;
+    margin: 0;
+}
+
+.section__balance--negative {
+    color: var(--color-down);
+}
+
+.section__text {
+    font-size: 24px;
+    font-weight: 700;
+    color: var(--color-text);
+    line-height: 1;
+    margin: 0;
+}
+
+.section__text--reversed {
+    text-align: right;
+    transform: rotateY(-180deg);
+}
+
+.section--column > .section__title {
+    flex: 0 0 auto;
+}
+
+.section__body {
+    box-sizing: border-box;
+    width: 100%;
+    height: auto;
+}
+
+.section--column > .section__body {
+    flex: 1 1 0;
+}
