+++ src/styles/tab.css
@@ -1,0 +1,41 @@
+.tab {
+    box-sizing: border-box;
+    position: relative;
+    display: flex;
+    width: 124px;
+    height: 100%;
+    border-top: 4px solid transparent;
+    border-bottom: 4px solid transparent;
+}
+
+.fleeca .tab--active {
+    border-bottom: 4px solid var(--color-fleeca);
+}
+
+.maze .tab--active {
+    border-bottom: 4px solid var(--color-maze);
+}
+
+.tab__label {
+    margin: auto;
+    font-size: 12px;
+    font-weight: 700;
+    text-transform: uppercase;
+    letter-spacing: 0.6px;
+    color: var(--color-text-alt);
+}
+
+.tab--active > .tab__label {
+    color: var(--color-text);
+}
+
+.tab__input {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    margin: 0;
+    appearance: none;
+    cursor: pointer;
+}
