+++ src/styles/top-app-bar.css
@@ -1,0 +1,63 @@
+.top-app-bar {
+    box-sizing: border-box;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    height: 72px;
+    padding-left: 24px;
+    padding-right: 24px;
+    background-color: var(--color-fg);
+    border-bottom: 1px solid var(--color-border);
+}
+
+.top-app-bar__section {
+    box-sizing: border-box;
+    flex: 1 1 0;
+}
+
+.top-app-bar__section--middle {
+    height: 100%;
+    display: flex;
+    flex-direction: row;
+}
+
+.top-app-bar__section--right {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    align-items: flex-start;
+}
+
+.top-app-bar__section--row {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+}
+
+.top-app-bar__logo {
+    box-sizing: border-box;
+    display: block;
+    width: 36px;
+    height: auto;
+    margin-right: 16px;
+    pointer-events: none;
+}
+
+.maze .top-app-bar__logo {
+    color: var(--color-maze);
+}
+
+.fleeca .top-app-bar__logo {
+    color: var(--color-fleeca);
+}
+
+.maze .top-app-bar__logo--fleeca {
+    display: none;
+}
+
+.fleeca .top-app-bar__logo--maze {
+    display: none;
+}
